Add a magic "fix everything" command
f80211 appears to wait for ...something to finish, but new jobs on
other GPU cores might still not read from the old core's cachelines
correctly. Previous Mali GPUs must've used a different mechanism for
messing up wifi when jobs fault.

0x233 waits for everything, so appears to be what we want to use. Here
I'm passing in a parameter of 0, but the blob also uses 0xffffe0 and
1. They don't seem to be as effective for some reason.