Magically fix things another way
the '8' means:

"Page coherence Inner shareable"

Inner shareable in this case means the GPU, as opposed to Outer
shareable which presumably includes the CPU, and System which is I
guess useful when sending commands directly to an nVME drive or
something like that. (This is all ARM/AArch64 terminology.)

So without this flag set, GPU cores will not check the various L1
caches of the other cores when allocating a cache line for memory,
If this is set, I guess it is equivalent to doing "dsb ish" after
every access. Maybe. Sorta.

The only case where this is a problem is read-modify-write (or
atomic!) workloads: Read accesses are fine because there are no
changes to get out of sync, and pure writes are fine for at least some
access methods because L1 caches appear to not be write-allocate. So
if only writes are done, they will go directly to L2 (or even main
memory?) without allocating a cache line, and the granularity of
accesses will be much less problematic.

I note that it is probably safe to not set the flag for framebuffer
contents, since IIRC v10 keeps using a single core for the whole of
each tile.