panfrost: Implement depth clamp for v9 and later
Finally we have conforming depth clamp support!

Although AMD_depth_clamp_separate specifies clamping in terms of near
and far planes, hardware uses "low" and "high" depth, so the bits have
to be reversed when a reversed depth range is in effect.

Note that setting minimum_z and maximum_z to ±INFINITY now causes the
entire polygon to be culled, though large finite values do not.