pan/base: Rewrite syncobj handling
It turns out that having syncobjs depend on each other in a tree
rather than being self-contained is a really stupid idea, and can
easily cause stack overflows, not to mention the memory wastage.

So flip them on their head--rather than jobs referencing syncobjs, the
syncobjs reference the jobs through a list of fences, which is a more
sane to do things.

  I realise that it would've been possible to keep the concept of
  fences but make job completion mark those fences as signaled, so
  then we can keep a roughly similar architecture without the problems
  from having a tree, but that isn't significantly better than how
  this patch does it.

One problem is that if a syncobj is not waited on for a long time, it
could accumulate a long list of fences, but that could be fixed by
calling kbase_syncobj_update occasionally when adding to the syncobj.

So now the callback system originally added for syncobjs is no longer
used for that purpose, but it *is* still used for BO freeing. This is
a problem--we no longer know what the most recently submitted syncobj
for a given queue is--but that can be fixed by adding a null
"sync_link" (which should probably be renamed) with no callback.

Another problem is what happens when an event slot is reused. How do
we know whether a syncobj is pointing to and old or new version of a
slot? The fix for that is simply to keep a linked list of all syncobjs
to iterate over.