panfrost: Execute fragment jobs on scoreboard slot 4
The cache clean has a mask of 0xf8, which means that cachelines only
allocated from slot 2 won't be written out to memory immediately.

Fix this by using slot 4, which is covered by the mask. I tried
changing the mask to 0xff, but that caused problems. But maybe that's
just because bit 0 shouldn't be set?

TODO: Rotate through scoreboards, rather than using a static
assignment.