panfrost: Fix Xwayland memory leak
If batches are being created but not submitted, memory was getting
leaked over time as events were never getting handled. Xwayland hit
this, because it was calling glFlush but not doing any rendering.

Also handle events in the panfrost_bo_create failure case, in case
there is some other usage pattern that hits this problem,
e.g. creating and destroying textures but never rendering.