panfrost: Handle race condition for seqnum values
It's possible for the seqnum value to reference an unsubmitted batch,
as dependencies are updated after the seqnum is incremented, but
before the batch is actually submitted. Handle this by instead waiting
on the previous job, assuming the worst case for what was there before
the seqnum was changed for the unsubmitted batch.